+++ mpapt-runtime/src/main/java/de/jensklingenberg/mpapt/common/MpAptProject.kt
@@ -10,22 +10,15 @@ import org.jetbrains.kotlin.codegen.extensions.ClassBuilderInterceptorExtension
 import org.jetbrains.kotlin.config.CompilerConfiguration
 import org.jetbrains.kotlin.extensions.StorageComponentContainerContributor
 import org.jetbrains.kotlin.js.translate.extensions.JsSyntheticTranslateExtension
-import org.jetbrains.kotlin.resolve.extensions.SyntheticResolveExtension
 
 class MpAptProject(abstractProcessor: AbstractProcessor, configuration: CompilerConfiguration) :
-        SyntheticResolveExtension by SyntheticResolveExtensionImpl(abstractProcessor, ClassParser(abstractProcessor)),
         IrGenerationExtension by NativeIrGenerationExtension(abstractProcessor),
         JsSyntheticTranslateExtension by JsSyntheticTranslateExtensionExt(abstractProcessor),
         ClassBuilderInterceptorExtension by ClassBuilderInterceptorExtensionImpl(abstractProcessor),
         StorageComponentContainerContributor by StorageComponentContainerContributorImpl(abstractProcessor)
-// CompilerConfigurationExtension by CompilerConfigurationExtensionExt(abstractProcessor),
-//  ExpressionCodegenExtension by ExpressionCodegenExtensionImpl(abstractProcessor)
-//  PreprocessedVirtualFileFactoryExtension by MetaPreprocessedVirtualFileFactoryExtension(abstractProcessor)
-//   AnalysisHandlerExtension by AnalysisHandlerExtensionImpl(abstractProcessor),
 {
     init {
         abstractProcessor.configuration = (configuration)
         abstractProcessor.onProcessingStarted()
-
     }
 }