+++ mpapt-runtime/src/main/java/de/jensklingenberg/mpapt/common/SyntheticResolveExtensionImpl.kt
@@ -1,7 +1,0 @@
-package de.jensklingenberg.mpapt.common
-
-import de.jensklingenberg.mpapt.model.AbstractProcessor
-import org.jetbrains.kotlin.resolve.extensions.SyntheticResolveExtension
-
-@Deprecated("This class is not needed anymore")
-class SyntheticResolveExtensionImpl(abstractProcessor: AbstractProcessor, classParser: ClassParser) : SyntheticResolveExtension