+++ docs/CHANGELOG.md
@@ -1,3 +1,16 @@
+v0.8.4
+========================================
+
+* Build with Kotlin 1.3.61
+
+* Fixed bug that produced a crash when annotating functions with expression body like
+ "fun functionWithExpressionBody() = 1 + 1" on Native (thx to @chippmann)
+ 
+
+* Removed SyntheticResolveExtensionImpl, you don't need add:
+ SyntheticResolveExtension.registerExtension(project, mpapt) anymore
+
+
 v0.8.3
 ========================================
 * You can now detect annotated expressions in functions